requirements.txt
attrs==20.3.0
blessed==1.17.6
botocore==1.20.59
chalice==1.22.4
click==7.1.2
inquirer==2.7.0
jmespath==0.10.0
mypy-extensions==0.4.3
PyJWT==2.0.1
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.4.1
readchar==2.0.1
redis==3.5.3
six==1.15.0
urllib3==1.26.4
wcwidth==0.2.5
